components/auth-service-backend-dev/auth-service-backend-dev/requirements.txt
aiosqlite==0.21.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
fastapi==0.128.7
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
asyncpg==0.31.0
psycopg2-binary==2.9.10
bcrypt==3.2.0
redis==7.2.1
